Fixes after update to Nuxt 3.1.0
