+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "nuxt-graphql-middleware",
-  "version": "3.1.0-beta.0",
+  "version": "3.1.0-beta.2",
   "description": "Module to perform GraphQL requests as a server middleware.",
   "repository": {
     "type": "git",
@@ -35,11 +35,11 @@
     "test:coverage": "vitest run --coverage"
   },
   "dependencies": {
-    "@graphql-codegen/cli": "^2.15.0",
-    "@graphql-codegen/schema-ast": "^2.5.1",
-    "@graphql-codegen/typescript": "^2.8.3",
+    "@graphql-codegen/cli": "^2.16.4",
+    "@graphql-codegen/schema-ast": "^2.6.1",
+    "@graphql-codegen/typescript": "^2.8.7",
     "@graphql-codegen/typescript-generic-sdk": "^3.0.4",
-    "@graphql-codegen/typescript-operations": "^2.5.8",
+    "@graphql-codegen/typescript-operations": "^2.5.12",
     "@graphql-fragment-import/lib": "^2.0.0",
     "@graphql-tools/utils": "^9.1.1",
     "@nuxt/kit": "^3.0.0",
@@ -49,21 +49,21 @@
   "devDependencies": {
     "@nuxt/module-builder": "^0.2.1",
     "@nuxt/schema": "^3.0.0",
-    "@nuxt/test-utils": "^3.0.0",
+    "@nuxt/test-utils": "^3.1.1",
     "@nuxtjs/eslint-config-typescript": "^12.0.0",
     "@types/capture-console": "^1.0.1",
     "@types/cli-table": "^0.3.1",
     "@types/inquirer": "^9.0.3",
-    "@vitest/coverage-c8": "^0.25.3",
+    "@vitest/coverage-c8": "^0.25.8",
     "cypress": "^11.2.0",
-    "eslint": "^8.29.0",
-    "eslint-config-prettier": "^8.5.0",
+    "eslint": "^8.33.0",
+    "eslint-config-prettier": "^8.6.0",
     "eslint-plugin-prettier": "^4.2.1",
     "jsdoc-to-markdown": "^8.0.0",
-    "nuxt": "^3.0.0",
-    "prettier": "^2.8.0",
+    "nuxt": "^3.1.1",
+    "prettier": "^2.8.3",
     "strip-ansi": "^7.0.1",
-    "typedoc": "^0.23.21",
+    "typedoc": "^0.23.24",
     "typedoc-plugin-markdown": "^3.14.0",
     "vitepress": "^1.0.0-alpha.30",
     "vitest": "^0.25.3"
