+++ playground/nuxt.config.ts
@@ -1,8 +1,7 @@
 import { defineNuxtConfig } from 'nuxt/config'
-import { GraphqlMiddlewareConfig } from '../src/types'
 import graphqlMiddlewareModule from './../src/module'
 
-const graphqlMiddleware: GraphqlMiddlewareConfig = {
+const graphqlMiddleware = {
   graphqlEndpoint: 'http://localhost:4000',
   downloadSchema: false,
   // serverFetchOptions: function () {},
