+++ src/runtime/serverOptions/defineGraphqlServerOptions.ts
@@ -1,0 +1,7 @@
+import { GraphqlMiddlewareServerOptions } from './../../types'
+
+export function defineGraphqlServerOptions(
+  options: GraphqlMiddlewareServerOptions,
+) {
+  return options
+}
