+++ src/runtime/serverOptions/index.ts
@@ -1,0 +1,3 @@
+import { defineGraphqlServerOptions } from './defineGraphqlServerOptions'
+
+export { defineGraphqlServerOptions }
