+++ src/types.ts
@@ -1,6 +1,5 @@
 import type { H3Event } from 'h3'
 import type { FetchOptions, FetchResponse, FetchError } from 'ofetch'
-import { TypeScriptDocumentsPluginConfig } from '@graphql-codegen/typescript-operations'
 import type { GraphQLError } from 'graphql'
 
 export type GraphqlMiddlewareGraphqlEndpointMethod = (
@@ -28,124 +27,6 @@ export type GraphqlMiddlewareOnServerErrorMethod = (
   operation?: string,
   operationName?: string,
 ) => any | Promise<any>
-
-export interface GraphqlMiddlewareConfig {
-  /**
-   * File glob patterns for the auto import feature.
-   *
-   * If left empty, no documents are auto imported.
-   *
-   * @default
-   * ```json
-   * ["**\/.{gql,graphql}", "!node_modules"]
-   * ```
-   *
-   * @example
-   * ```ts
-   * // Load .graphql files from pages folder and from a node_modules dependency.
-   * const autoImportPatterns = [
-   *   './pages/**\/*.graphql',
-   *   'node_modules/my_library/dist/**\/*.graphql'
-   * ]
-   * ```
-   */
-  autoImportPatterns?: string[]
-
-  /**
-   * Additional raw documents to include.
-   *
-   * Useful if for example you need to generate queries during build time.
-   *
-   * @default []
-   *
-   * @example
-   * ```ts
-   * const documents = [`
-   *   query myQuery {
-   *     articles {
-   *       title
-   *       id
-   *     }
-   *   }`,
-   *   ...getGeneratedDocuments()
-   * ]
-   * ```
-   */
-  documents?: string[]
-
-  /**
-   * Wether the useGraphqlQuery, useGraphqlMutation and useGraphqlState
-   * composables should be included.
-   *
-   * @default ```ts
-   * true
-   * ```
-   */
-  includeComposables?: boolean
-
-  /**
-   * Enable detailled debugging messages.
-   *
-   * @default false
-   */
-  debug?: boolean
-
-  /**
-   * The URL of the GraphQL server.
-   *
-   * For the runtime execution you can provide a method that determines the endpoint
-   * during runtime. See the app/graphqlMiddleware.serverOptions.ts documentation
-   * for more information.
-   */
-  graphqlEndpoint: string
-
-  /**
-   * Download the GraphQL schema and store it on disk.
-   *
-   * @default true
-   */
-  downloadSchema?: boolean
-
-  /**
-   * The prefix for the server route.
-   *
-   * @default ```ts
-   * "/api/graphql_middleware"
-   * ```
-   */
-  serverApiPrefix?: string
-
-  /**
-   * Path to the GraphQL schema file.
-   *
-   * If `downloadSchema` is `true`, the downloaded schema is written to this specified path.
-   * If `downloadSchema` is `false`, this file must be present in order to generate types.
-   *
-   * @default './schema.graphql'
-   */
-  schemaPath?: string
-
-  /**
-   * These options are passed to the graphql-codegen method when generating the operations types.
-   *
-   * {@link https://www.the-guild.dev/graphql/codegen/plugins/typescript/typescript-operations}
-   * @default
-   * ```ts
-   * const codegenConfig = {
-   *   exportFragmentSpreadSubTypes: true,
-   *   preResolveTypes: true,
-   *   skipTypeNameForRoot: true,
-   *   skipTypename: true,
-   *   useTypeImports: true,
-   *   onlyOperationTypes: true,
-   *   namingConvention: {
-   *     enumValues: 'change-case-all#upperCaseFirst',
-   *   },
-   * }
-   * ```
-   */
-  codegenConfig?: TypeScriptDocumentsPluginConfig
-}
 
 /**
  * Configuration options during runtime.
